requirements.txt
torch
librosa
transformers
hydra-core
omegaconf
fire